+++ clients/derivatives-trading-portfolio-margin/src/rest-api/types/new-um-algo-order-response.ts
@@ -114,6 +114,12 @@ export interface NewUmAlgoOrderResponse {
      * @type {boolean}
      * @memberof NewUmAlgoOrderResponse
      */
+    closePosition?: boolean;
+    /**
+     *
+     * @type {boolean}
+     * @memberof NewUmAlgoOrderResponse
+     */
     priceProtect?: boolean;
     /**
      *